+++ packages/agtree/scripts/.eslintrc.cjs
@@ -1,0 +1,10 @@
+/**
+ * @file ESLint configuration for scripts.
+ */
+
+module.exports = {
+    rules: {
+        'import/no-extraneous-dependencies': 0,
+        'no-console': 0,
+    },
+};