+++ packages/agtree/scripts/check-compatibility-tables.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-console, import/no-extraneous-dependencies */
 /**
  * @file Very simple script to validate the compatibility tables.
  */